requirements.txt
# FastAPI and ASGI server
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Supabase integration
supabase==2.0.0

# Data validation and serialization
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# HTTP client for external APIs
httpx==0.24.1

# Environment variables
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
aiosqlite==0.19.0

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Logging
structlog==23.2.0

# Email & Scheduling Dependencies
fastapi-mail==1.4.1
apscheduler==3.10.4
jinja2==3.1.2
email-validator==2.1.0
PyJWT==2.8.0